Rendering Cart Data and Handling Cart Totals
Rendering the cart data to show the user the added products, including the products':
 - names
 - images
 - quantities
 - prices
The condition to check whether the user is logged in or a guest was also added to views.py.

Additionally, the following methods were created to dynamically show data to the user:
1. get_total - show the user the total cost of each product
2. get_cart_total - the total cost of the order
3. get_cart_items - the total no. of items being ordered
This ensures that as new products are added or existing products' quantities are updated, the prices & total quantities are dynamically updated too.