Creating EventHandler for 'Add to Cart'
Creating a javascript file and linking it to my HTML code via main.html.

Creating a class called update-cart in store.html to recognize and store product ID and 'add' action when the 'add to cart' button is clicked.

Then, creating an event handler in cart.js to recognize this clicking of 'add to cart' by user, and looping through the clicked buttons to get the products clicked on.

Also adding functionality to check whether the user adding the products is logged in or not - which will help when adding items to User cart, or guest cart.