Sending 'Add to Cart' Data to View
Using a fetch api in the cart.js file and connecting it to the updateItem view in views.py to pass a POST request and upload the added product to the updateItem view.

Also resolving the problem of a missing CSRFToken when passing POST requests by using a function from the Django documentation to manually generate a token.