Updating Items in Cart
Adding functionality to the up & down arrow keys in the Cart to ensure that product quantities can be updated by clicking them too.

NOTE: The process of updating items is currently inefficient, as every time a product quantity is updated, the page reloads to reflect the change. This process can become highly inefficient when the website server is handling multiple requests simultaneously.