Creating/Retrieving Cart from Cookies
Using the getCookie function to retrieve the cart variable from the site cookies. If the cart does not exist, then the function creates it.

In the cookies, the cart is stored as a string (hence being passed via JSON.stringify into document.cookies). However, in the code, it will be dealt with as a JSON object, and is thus parsed when stored into the variable of cart.

We are going to store guest orders in this JSON cart variable, which will subsequently be updated to the site cookies. This is so that whenever the guest reloads the site (after however many days), the cart is still stored in their local site cookies.