Updating Items to Cart Cookie
Writing a function to updating the product ID and quantity of a product in the JSON cart variable when the 'add to cart' button is clicked.

Uses a nested JSON object to track the product added, and the quantity of said product in the cart.

The updated JSON type 'cart' variable is then published to the cookies again to be stored in the backend, so it can be saved.