Building Cart Items
Iterating through the cart dictionary and adding the relevant product details, quantity, and total to an item dictionary to create an OrderItem object. This is done to ensure that in cart.html, when we run the for loop of {% for item in items %}, the relevant attributes of id, imageURL, price etc. that we query are available in the item dictionary.