Handling Product Removals by Admin
Currently, the products being added to cart are stored in site cookies, independent of the server database. Thus, if a product is added to cart by the user, but is then later removed by the admin; that'll throw up an error for the user.

To prevent that, we encase all the queries in querying_data in a try-except clause to catch and pass such exceptions.

Also putting the entire querying_data into a utils.py file and importing it into views.py to use. This was done simply to make the code cleaner and more organised.