Creating Login and Logout Pages
I next created the Login and Logout pages. To create the Login page, I:
1. Created a login url under urls.py

2. Invoked Django's AuthenticationForm under the login_request view, which is used to check whether a User record exists

3. Added an HTML page to provide a fronted to log in.

When creating the HTML page, I created a base html file (base.html) for the login/sign up pages to inherit from, to provide standardisation.

Then, to create the logout page, I simply called upon Django's "Logout" function, which logs out the customer.

Furthermore, I added navigation bars on the sign up & login pages to make functionality between sign up, login, and logout easier. I also modified the main navigation bar on the store's homepage to give a 'login' option if the customer is not logged in, and a 'logout' option if they are already logged in.