Changing ShippingAddress & Order Dependencies
I changed the relationship between ShippingAddress and Order so that now, they have a 1-Many relationship (i.e., many orders can have same shipping address). To do so, I placed a foreign key for a shipping address in the Order table.
I also updated the post processing to add the shipping address ID to the required field in the Order record before saving.