Change Date Ordered field of Order
I changed the date ordered field of order to save the date at the moment that the order is placed, rather than when the cart is created. This ensures that even if a customer leaves a cart open for a long period of time, the recorded time of order placement is only when their order successfully completes.